refactor: move the tx display cache code to a singleton service
