+++ wallet/res/values/strings-extra.xml
@@ -73,6 +73,7 @@
     <string name="history_refresh_dialog_title">Refresh Transaction List</string>
     <string name="history_refresh_dialog_message">This will clear the transaction display cache and rebuild it from scratch. The list will be temporarily unavailable while it rebuilds.</string>
     <string name="history_refresh_dialog_confirm">Refresh</string>
+    <string name="history_refresh_complete">Transaction list refreshed</string>
     <string name="more_title">More</string>
     <string name="settings_title">Settings</string>
     <string name="security_title">Security</string>
